# Conflicts:
#	src/main/java/com/example/project11/Controllers/SeparateQuestions.java
#	src/main/resources/com/example/project11/separateQuestions.fxml

+++ src/main/java/com/example/project11/ProjectInfo/STEP2/GraduatePassingGradePredictionThreshold.java
@@ -10,7 +10,7 @@ import java.util.Map;
 
 public class GraduatePassingGradePredictionThreshold {
     // Creating a hash map
-    private Map<String, Double> coursePassingMap;
+    private Map<String, Integer> coursePassingMap;
     private double[][] allStudents ;
 
     public GraduatePassingGradePredictionThreshold(double[][] allstudents) {
@@ -24,7 +24,7 @@ public class GraduatePassingGradePredictionThreshold {
             double[] passingPercentages = calculatePassingPercentagesArray(allStudents);
 
             for (int i = 0; i < passingPercentages.length; i++) {
-                coursePassingMap.put(Integer.toString(i), passingPercentages[i]); // Course numbers start at 1
+                coursePassingMap.put(Integer.toString(i), (int)passingPercentages[i]); // Course numbers start at 1
             }
 
 
@@ -51,10 +51,8 @@ public class GraduatePassingGradePredictionThreshold {
 
 
     // a getter for the hash map
-    public Map<String, Double> getCoursePassingMap() {
+    public Map<String, Integer> getCoursePassingMap() {
         calculatePassingPercentagesArray();
-
-
         return coursePassingMap;
     }
 }
